whisper_app/requirements.txt
Flask
openai-whisper @ git+https://github.com/openai/whisper.git
torch
ffmpeg-python
